+++ waspc/examples/todoApp/src/Todo.tsx
@@ -1,5 +1,4 @@
 import { Link } from 'wasp/client/router'
-import { type Task } from 'wasp/entities'
 
 import {
   useAction,
@@ -10,23 +9,22 @@ import {
   toggleAllTasks,
   useQuery,
   getTasks,
-  getTask,
-  getDate,
-  getAnythingAuth,
 } from 'wasp/client/operations'
 
 import React, { useState, FormEventHandler, ChangeEventHandler } from 'react'
+import { getEmail } from 'wasp/auth'
 
 type NonEmptyArray<T> = [T, ...T[]]
 
+type TaskWithUser = Awaited<ReturnType<typeof getTasks>>[number]
+
 export function areThereAnyTasks(
-  tasks: Task[] | undefined
-): tasks is NonEmptyArray<Task> {
+  tasks: TaskWithUser[] | undefined
+): tasks is NonEmptyArray<TaskWithUser> {
   return !!(tasks && tasks.length > 0)
 }
 
 const Todo = () => {
-  // logAll()
   const { data: tasks, isError, error: tasksError } = useQuery(getTasks)
 
   const TasksError = () => {
@@ -59,7 +57,7 @@ const Todo = () => {
   )
 }
 
-const Footer = ({ tasks }: { tasks: NonEmptyArray<Task> }) => {
+const Footer = ({ tasks }: { tasks: NonEmptyArray<TaskWithUser> }) => {
   const numCompletedTasks = tasks.filter((t) => t.isDone).length
   const numUncompletedTasks = tasks.filter((t) => !t.isDone).length
 
@@ -87,7 +85,7 @@ const Footer = ({ tasks }: { tasks: NonEmptyArray<Task> }) => {
   )
 }
 
-const Tasks = ({ tasks }: { tasks: NonEmptyArray<Task> }) => {
+const Tasks = ({ tasks }: { tasks: NonEmptyArray<TaskWithUser> }) => {
   return (
     <div>
       <table className="border-separate border-spacing-2">
@@ -101,9 +99,9 @@ const Tasks = ({ tasks }: { tasks: NonEmptyArray<Task> }) => {
   )
 }
 
-type UpdateTaskIsDonePayload = Pick<Task, 'id' | 'isDone'>
+type UpdateTaskIsDonePayload = Pick<TaskWithUser, 'id' | 'isDone'>
 
-const TaskView = ({ task }: { task: Task }) => {
+const TaskView = ({ task }: { task: TaskWithUser }) => {
   const updateTaskIsDoneOptimistically = useAction(updateTaskIsDone, {
     optimisticUpdates: [
       {
@@ -118,24 +116,7 @@ const TaskView = ({ task }: { task: Task }) => {
             )
           }
         },
-      } as OptimisticUpdateDefinition<UpdateTaskIsDonePayload, Task[]>,
-      {
-        getQuerySpecifier: () => [getTask, { id: task.id }],
-        // This query's cache should should never be emtpy
-        updateQuery: ({ isDone }, oldTask) => {
-          if (oldTask === undefined) {
-            // Cache is empty (e.g., the user has not yet opened the task's
-            // dedicated page.
-            // Passes the type checker because of the assertion Returning
-            // undefined should be properly supported and not a hack, see
-            // https://github.com/wasp-lang/wasp/issues/2017
-            return undefined
-          } else {
-            const result = { ...oldTask!, isDone }
-            return result
-          }
-        },
-      } as OptimisticUpdateDefinition<UpdateTaskIsDonePayload, Task>,
+      } as OptimisticUpdateDefinition<UpdateTaskIsDonePayload, TaskWithUser[]>,
     ],
   })
   const handleTaskIsDoneChange: ChangeEventHandler<HTMLInputElement> = async (
@@ -151,6 +132,8 @@ const TaskView = ({ task }: { task: Task }) => {
     }
   }
 
+  const email = getEmail(task.user)
+
   return (
     <tr>
       <td>
@@ -164,7 +147,7 @@ const TaskView = ({ task }: { task: Task }) => {
       </td>
       <td>
         <Link to="/task/:id" params={{ id: task.id }}>
-          {task.description}
+          {task.description} {email && `by ${email}`}
         </Link>
       </td>
     </tr>
@@ -179,21 +162,26 @@ const NewTaskForm = () => {
       {
         getQuerySpecifier: () => [getTasks],
         updateQuery: (newTask, oldTasks) => {
+          const newTaskWithUser = {
+            ...newTask,
+            user: {},
+          } as TaskWithUser
+
           if (oldTasks === undefined) {
             // cache is empty
-            return [newTask as Task]
+            return [newTaskWithUser]
           } else {
-            return [...oldTasks, newTask as Task]
+            return [...oldTasks, newTaskWithUser]
           }
         },
       } as OptimisticUpdateDefinition<
-        Pick<Task, 'isDone' | 'description'>,
-        Task[]
+        Pick<TaskWithUser, 'isDone' | 'description'>,
+        TaskWithUser[]
       >,
     ],
   })
 
-  const createNewTask = async (description: Task['description']) => {
+  const createNewTask = async (description: TaskWithUser['description']) => {
     const task = { isDone: false, description }
     await createTaskFn(task)
   }
@@ -241,26 +229,6 @@ const ToggleAllTasksButton = ({ disabled }: { disabled: boolean }) => {
       ✓
     </button>
   )
-}
-
-// Use this function to test calling actions directly
-async function logAll() {
-  const tasks = await getTasks()
-  console.info('Got tasks:', tasks)
-
-  const someId = tasks.map((task) => task.id).find((id) => id)
-  if (!someId) {
-    console.info('No tasks found')
-  } else {
-    const task = await getTask({ id: someId })
-    console.info(`Got task with id ${someId}`, task)
-  }
-
-  const date = await getDate()
-  console.info('Got date:', date)
-
-  const anything = await getAnythingAuth()
-  console.info('Got anything:', anything)
 }
 
 export default Todo