+++ waspc/packages/deploy/.eslintrc.cjs
@@ -1,83 +1,57 @@
 module.exports = {
-  "env": {
-    "es2020": true,
-    "node": true
+  env: {
+    es2020: true,
+    node: true,
   },
-  "root": true,
-  "extends": [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended"
-  ],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "ecmaVersion": 11,
-    "sourceType": "module"
+  root: true,
+  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    ecmaVersion: 11,
+    sourceType: 'module',
   },
-  "plugins": [
-    "@typescript-eslint"
-  ],
-  "rules": {
-    "indent": [
-      "error",
-      "tab"
-    ],
-    "linebreak-style": [
-      "error",
-      "unix"
-    ],
-    "quotes": [
-      "error",
-      "single"
-    ],
-    "eol-last": [
-      "error",
-      "always"
-    ],
-    "no-multiple-empty-lines": [
-      "error",
+  plugins: ['@typescript-eslint'],
+  rules: {
+    indent: ['error', 'space'],
+    'linebreak-style': ['error', 'unix'],
+    quotes: ['error', 'single'],
+    'eol-last': ['error', 'always'],
+    'no-multiple-empty-lines': [
+      'error',
       {
-        "max": 2,
-        "maxEOF": 1
-      }
+        max: 2,
+        maxEOF: 1,
+      },
     ],
-    "comma-spacing": [
-      "error",
-      { "before": false, "after": true }
-    ],
-    "space-before-function-paren": ["error", {
-      "anonymous": "always",
-      "named": "never",
-      "asyncArrow": "always"
-    }],
-    "comma-dangle": [
-      "error",
-      "always-multiline"
-    ],
-    "object-curly-spacing": [
-      "error",
-      "always"
-    ],
-    "padding-line-between-statements": [
-      "error",
-      { "blankLine": "always", "prev": "function", "next": "function" },
-      { "blankLine": "always", "prev": "function", "next": "export" },
-      { "blankLine": "always", "prev": "export", "next": "function" },
-      { "blankLine": "always", "prev": "export", "next": "export" }
-    ],
-    "no-duplicate-imports": "error",
-    "@typescript-eslint/semi": [
-      "error",
-      "always"
-    ],
-    "@typescript-eslint/member-delimiter-style": [
-      "error",
+    'comma-spacing': ['error', { before: false, after: true }],
+    'space-before-function-paren': [
+      'error',
       {
-        "multiline": {
-          "delimiter": "semi",
-          "requireLast": true
-        }
-      }
+        anonymous: 'always',
+        named: 'never',
+        asyncArrow: 'always',
+      },
     ],
-    "@typescript-eslint/explicit-module-boundary-types": "error"
-  }
-}
+    'comma-dangle': ['error', 'always-multiline'],
+    'object-curly-spacing': ['error', 'always'],
+    'padding-line-between-statements': [
+      'error',
+      { blankLine: 'always', prev: 'function', next: 'function' },
+      { blankLine: 'always', prev: 'function', next: 'export' },
+      { blankLine: 'always', prev: 'export', next: 'function' },
+      { blankLine: 'always', prev: 'export', next: 'export' },
+    ],
+    'no-duplicate-imports': 'error',
+    '@typescript-eslint/semi': ['error', 'always'],
+    '@typescript-eslint/member-delimiter-style': [
+      'error',
+      {
+        multiline: {
+          delimiter: 'semi',
+          requireLast: true,
+        },
+      },
+    ],
+    '@typescript-eslint/explicit-module-boundary-types': 'error',
+  },
+};