+++ waspc/src/Wasp/Generator/Valid/TsConfig.hs
@@ -78,7 +78,7 @@ validateFieldValue fullyQualifiedFieldName expectedValue actualValue =
         unwords
           [ "Invalid value for the",
             "\"" ++ show fullyQualifiedFieldName ++ "\"",
-            "field in tsconfig.json, you must set it to:",
+            "field in TS config, you must set it to:",
             showAsJsValue expected ++ "."
           ]
 
@@ -87,7 +87,7 @@ validateFieldValue fullyQualifiedFieldName expectedValue actualValue =
         unwords
           [ "The",
             "\"" ++ show fullyQualifiedFieldName ++ "\"",
-            "field in tsconfig.json must be unset."
+            "field in TS Config must be left unspecified."
           ]
 
     makeMissingFieldErrorMessage expected =
@@ -95,6 +95,6 @@ validateFieldValue fullyQualifiedFieldName expectedValue actualValue =
         unwords
           [ "The",
             "\"" ++ show fullyQualifiedFieldName ++ "\"",
-            "field is missing in tsconfig.json, you must set it to:",
+            "field is missing in TS config, you must set it to:",
             showAsJsValue expected ++ "."
           ]