+++ waspc/src/Wasp/Project/ExternalConfig/TsConfig.hs
@@ -7,13 +7,14 @@ import Control.Arrow (left)
 import Control.Monad.Except (ExceptT (ExceptT), runExceptT)
 import qualified Data.ByteString.Lazy.UTF8 as BS
 import Data.Either.Extra (maybeToEither)
-import StrongPath (Abs, Dir, File, Path', Rel, toFilePath)
+import StrongPath (Abs, Dir, File, Path', Rel, basename, fromRelFile, toFilePath)
 import qualified Wasp.ExternalConfig.TsConfig as T
 import Wasp.Project.Common
   ( SrcTsConfigFile,
     WaspProjectDir,
     findFileInWaspProjectDir,
   )
+import Wasp.Util (indent)
 import qualified Wasp.Util.IO as IOUtil
 import Wasp.Util.Json (parseJsonWithComments)
 
@@ -32,4 +33,7 @@ readTsConfigFile :: Path' Abs (File SrcTsConfigFile) -> IO (Either String T.TsCo
 readTsConfigFile tsConfigFile = do
   tsConfigContent <- IOUtil.readFileBytes tsConfigFile
   parseResult <- parseJsonWithComments . BS.toString $ tsConfigContent
-  return $ left ("Failed to parse tsconfig file: " ++) parseResult
+  return $ left ((errorMessagePrefix ++) . indent 2) parseResult
+  where
+    errorMessagePrefix = "Failed to parse '" ++ baseTsConfigFilePath ++ "':\n"
+    baseTsConfigFilePath = fromRelFile (basename tsConfigFile)